requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
langchain
langchain-google-genai
langchain-community
langchain-chroma
chromadb
pypdf
tiktoken
